Add test for bootstrap
