+++ tests/common.py
@@ -9,6 +9,7 @@ from datetime import timedelta
 from unittest import mock
 
 import homeassistant as ha
+import homeassistant.util.location as location_util
 import homeassistant.util.dt as dt_util
 from homeassistant.helpers.entity import ToggleEntity
 from homeassistant.const import (
@@ -38,6 +39,23 @@ def get_test_home_assistant(num_threads=None):
     hass.config.longitude = -117.22743
 
     return hass
+
+
+def mock_detect_location_info():
+    """ Mock implementation of util.detect_location_info. """
+    return location_util.LocationInfo(
+        ip='1.1.1.1',
+        country_code='US',
+        country_name='United States',
+        region_code='CA',
+        region_name='California',
+        city='San Diego',
+        zip_code='92122',
+        time_zone='America/Los_Angeles',
+        latitude='2.0',
+        longitude='1.0',
+        use_fahrenheit=True,
+    )
 
 
 def mock_service(hass, domain, service):
